Compare text by UTF-8 byte order in the regex engine
